09-26-2020

+++ src/app/web/services/cart.service.ts
@@ -2,15 +2,15 @@ import {Injectable} from '@angular/core';
 import {HttpClient} from '@angular/common/http';
 import {BASE_URL} from '../../config/config';
 import {map} from 'rxjs/operators';
-import {Observable} from 'rxjs';
+import {combineLatest, Observable, zip} from 'rxjs';
 import {CartInterface} from '../interfaces/cart_item.interface';
 import {WebDataService} from './web-data.service';
 import {ItemModel} from '../models/item.model';
 import {Store} from '@ngrx/store';
 import {AppState} from '../store/app.reducer';
 import {UserModel} from '../models/user.model';
-import {ItemInterface} from '../interfaces/item.interface';
 import {SetStoreCartAction} from '../store/actions/cart.actions';
+import {WebRuc} from '../types/types';
 
 @Injectable({
   providedIn: 'root'
@@ -67,14 +67,51 @@ export class CartService {
       amount,
       code: item.code
     };
+
+    if (!this.user) {
+      return;
+    }
+
     return this.httpClient.post(`${BASE_URL}/cart/${this.user.id}/${item.ruc.slice(0, 2)}`,
       payload, {headers: this.dataService.headers()})
       .pipe(
-        map((res) => {
-          console.log(res);
-          return `Item agregado al carrito`;
-        })
+        map(() => `Item agregado al carrito`)
       );
+  }
+
+
+  addLocalItem(item: ItemModel, amount: number): Observable<string> {
+
+    return new Observable(subscriber => {
+
+      const localItems = this.dataService.getLocalCart();
+
+      let exits = false;
+      let message = 'Item Agregado';
+
+      localItems.forEach((litem: any) => {
+        if (litem.code === item.code) {
+          litem.amount += amount;
+          message = 'Item actualizado';
+          exits = true;
+        }
+      });
+
+      if (!exits) {
+        localItems.push({
+          prefix: item.ruc.slice(0, 2),
+          code: item.code,
+          amount
+        });
+      }
+
+      this.dataService.setCart(localItems);
+
+      setTimeout(() => {
+        subscriber.next(message);
+      }, 500);
+    });
+
   }
 
   updateItemCart(item: ItemModel, amount: number): Observable<string> {
@@ -82,6 +119,9 @@ export class CartService {
       amount,
       code: item.code
     };
+    if (!this.user) {
+      return;
+    }
     return this.httpClient.put(`${BASE_URL}/cart/${this.user.id}/${item.ruc.slice(0, 2)}`,
       payload, {headers: this.dataService.headers()})
       .pipe(
@@ -93,11 +133,42 @@ export class CartService {
   }
 
   removeItem(item: CartInterface): Observable<any> {
+    if (!this.user) {
+      return;
+    }
     return this.httpClient.delete(`${BASE_URL}/cart/${this.user.id}/${item.code}`, {
       headers: this.dataService.headers()
     })
       .pipe(
         map((res: any) => res.message)
       );
+  }
+
+  syncCart(): Observable<any> {
+
+    const items = this.dataService.getLocalCart();
+    this.dataService.dropCart();
+
+    const observables: Observable<any>[] = [];
+    if (items.length > 0) {
+
+      items.forEach((item) => {
+
+        if (item.prefix === '10' || item.prefix === '20') {
+          const model: any = {
+            code: item.code,
+            ruc: (item.prefix === '10') ? WebRuc.ROGER : WebRuc.GINDAR
+          };
+
+          observables.push(this.addItemCart(model, item.amount));
+        }
+
+
+      });
+      return combineLatest(observables);
+    } else {
+      return new Observable<any>();
+    }
+
   }
 }
